requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.19.0
python-dotenv==1.0.1
datagen-python-sdk>=0.1.1
